Added a plugin system, moved several format options into a plugin.
Introduced a plugin system similar to jQuery.fn.extend, which takes an
object as a paramter with the name of the function to override or create,
and the method to use instead.
Ex:
  //Plugin
  var plugin = {
    name: "MyPlugin",
    method: function() { console.log("my plugin") }
  }

  //Usage
  dayjs().extend(plugin);
  dayjs().MyPlugin();

Any plugins loaded will overwrite the global dayjs object (via protoype),
which will in turn cause any new dayjs instaces to use the new plugin method.

The number of additional format options was reduced to h/hh and a/A
for 12 hour time and AM/PM.