Update foramts and fix grammar and logic
This is the same update of formats that has been done about a month ago
on Moment.js.

Regarding the grammar, there was no logic for different grammar cases
for the nouns. This is added by adding very precise logic for
determining grammar cases by the `number`, `isFuture` and
`withoutPreffix` parameters. The same update has been submitted to
Moment.js.