add meridiem function to ku locale
Kurdish (ku) locale has it's own form of meridiem texts which are:
`AM` -> `پ.ن`
`PM` -> `د.ن`
The logic of determining which should be chosen is identical
to the main logic from the library, the only change is localized
strings for each case.
Also tests are added to the corresponding locale to ensure it
gets formatted correctly.