Correcting some false orders in the formats
Hi, dayjs is great. I have used it in my recent project. Thanks. But there is some formats which are incorrect in Tibetan. 
like, In Tibetan its, ལོ་ 6 (means 6 years) rather than 6 ལོ ( which sounds like years 6 in English ). 

Here are some correct formats that we are able to do. Thanks again! Cheers! :)