fix: arguments is of type object, but parse expects an array
CustomParseFormat.parse expects a string or an array as args parameter of
the cfg. When called recursively (via 'd.apply') dayjs added 'arguments'
which is of type object as cfg.args and therefore parse cannot access
e.g. the format as 'args[1]'.
The rest operator changes arguments to an array. And if dayjs gets a
config object with an args property as the 2nd parameter, we do not have
to add an args property, as everything is already in place.