fix: just using c parameter creates self reference
When the 2nd paramater of dayjs() is a config object, just assigning
'arguments' to the args property will modify the original 'c' parameter
and thus create a self referencing property (does not change functionality,
but should be avoided).
As 'arguments' is of type 'object', we must handle the case 'dayjs() with
multiple formats array, but without utc' by taking care of 'c' of type
array.