feat: make add / subtract with decimal values behave like moment
BREAKING CHANGE: in dayjs 1.0: only days and weeks are rounded to the
nearest integer.
In moment: days and months are rounded to the nearest integer;
Weeks, quarters, and years are converted to days or months, and then
rounded to the nearest integer.
Negative numbers are rounded as positive numbers with the sign added to
the result.