fix: customParseFormat: always use the index % 12
When the months/monthsShort locale date is split
into standalone and format parts the matchIndex
will be 12n + 11 for december, which ends up as 0
if one is added before the modulo, which results
in e.g. 0 || 23 == 23, interpreted by dayjs as dec
next year, since it's invalid to assume that the
original matchIndex is <= 12.

Instead do the modulo before adding one and adjust
the bounds check.

Fixes #2249